# Implement performance optimisation for string literal matching