@AntonLydike 2024-07-30
I will benchmark this tonight against some real-world workloads and report numbers