@AntonLydike 2024-08-02
It's hard to tell if this actually speeds anything up in practice, MLIRs benchmark suite goes from a geomean of `12.66` between 6 runs, to `12.49`, which is almost within error. xDSLs tests *increase* in duration from `8.06` to `8.25`, but these runs have even bigger error bars.

Further digging needed here.